requirements.txt
fastapi>=0.135.0
uvicorn>=0.41.0
aiosqlite>=0.20.0
PyJWT>=2.9.0
python-multipart>=0.0.18
pytest>=8.0.0
httpx>=0.27.0
